python-interpreter/packages/afm-core/pyproject.toml
[project]
name = "afm-core"
version = "0.1.3"
description = "AFM (Agent-Flavored Markdown) core: parser, CLI, protocols, and interfaces"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
]
license = "Apache-2.0"
requires-python = ">=3.11"
urls = { Repository = "https://github.com/wso2/reference-implementations-afm" }
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jsonschema>=4.26.0",
    "textual>=2.1.0",
    "fastapi>=0.128.1",
    "httpx>=0.28.1",
    "uvicorn>=0.40.0",
    "platformdirs>=4.0",
    "packaging>=24.0",
    "rich>=14.3.2",
]

[project.scripts]
afm = "afm.cli:cli"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "afm"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
